squash: script to populate starter buffer done!
